Minor changes, fixing the data type so it is consisten with the rest of the project.

+++ Second Project/listGraph.hpp
@@ -9,13 +9,13 @@ class List_Graph {
 private:
     std::vector<Vertex> nodes;
     std::unordered_map<Vertex, int> idx;
-    std::vector<std::vector<std::pair<int, int>>> adj_list; // (indeks s¹siada, waga)
+    std::vector<std::vector<std::pair<int, double>>> adj_list; // (indeks s¹siada, waga)
 public:
     List_Graph(const std::vector<Vertex>& nodes);
-    void addEdge(const Vertex& u, const Vertex& v, int weight);
-    std::vector<std::pair<Vertex, int>> neighbours(const Vertex& u) const;
+    void addEdge(const Vertex& u, const Vertex& v, double weight);
+    std::vector<std::pair<Vertex, double>> neighbours(const Vertex& u) const;
     bool hasEdge(const Vertex& u, const Vertex& v) const;
-    int getWeight(const Vertex& u, const Vertex& v) const;
+    double getWeight(const Vertex& u, const Vertex& v) const;
 };
 
 template<typename Vertex>
@@ -28,7 +28,7 @@ List_Graph<Vertex>::List_Graph(const std::vector<Vertex>& nodes) : nodes(nodes) 
 }
 
 template<typename Vertex>
-void List_Graph<Vertex>::addEdge(const Vertex& u, const Vertex& v, int weight) {
+void List_Graph<Vertex>::addEdge(const Vertex& u, const Vertex& v, double weight) {
     int iu = idx.at(u);
     int iv = idx.at(v);
     adj_list[iu].emplace_back(iv, weight);
@@ -36,9 +36,9 @@ void List_Graph<Vertex>::addEdge(const Vertex& u, const Vertex& v, int weight) {
 }
 
 template<typename Vertex>
-std::vector<std::pair<Vertex, int>> List_Graph<Vertex>::neighbours(const Vertex& u) const {
+std::vector<std::pair<Vertex, double>> List_Graph<Vertex>::neighbours(const Vertex& u) const {
     int iu = idx.at(u);
-    std::vector<std::pair<Vertex, int>> out;
+    std::vector<std::pair<Vertex, double>> out;
     for (const auto& p : adj_list[iu]) {
         out.emplace_back(nodes[p.first], p.second);
     }
@@ -59,11 +59,11 @@ bool List_Graph<Vertex>::hasEdge(const Vertex& u, const Vertex& v) const {
 }
 
 template<typename Vertex>
-int List_Graph<Vertex>::getWeight(const Vertex& u, const Vertex& v) const {
+double List_Graph<Vertex>::getWeight(const Vertex& u, const Vertex& v) const {
     int iu = idx.at(u);
     int iv = idx.at(v);
     for (const auto& p : adj_list[iu]) {
         if (p.first == iv) return p.second;
     }
-    return std::numeric_limits<int>::infinity(); // brak krawêdzi
+    return std::numeric_limits<double>::infinity(); // brak krawêdzi
 }
